+++ sdks/java/container/boot.go
@@ -105,8 +105,13 @@ func main() {
 		logger.Fatalf(ctx, "Failed to convert pipeline options: %v", err)
 	}
 
+	po, err := tools.ParseOptionsFromProto(info.GetPipelineOptions())
+	if err != nil {
+		logger.Fatalf(ctx, "Failed to parse pipeline options: %v", err)
+	}
+
 	// Inject artifact validation enabled state into context
-	ctx = artifact.WithArtifactValidation(ctx, !artifact.HasExperiment(info.GetPipelineOptions(), "disable_staged_file_integrity_checks"))
+	ctx = artifact.WithArtifactValidation(ctx, !po.HasExperiment("disable_staged_file_integrity_checks"))
 
 	// (2) Retrieve the staged user jars. We ignore any disk limit,
 	// because the staged jars are mandatory.